drawing lines implemented

+++ app/src/main/java/aa/engine/helpers/Position.kt
@@ -1,16 +1,16 @@
 package aa.engine.helpers
 
-class Position(private var x: Int, private var y: Int) {
+class Position(private var x: Float, private var y: Float) {
 
-    fun getX(): Int {
+    fun getX(): Float {
         return this.x;
     }
 
-    fun getY(): Int {
+    fun getY(): Float {
         return this.y;
     }
 
-    fun setPosition(x: Int, y: Int){
+    fun setPosition(x: Float, y: Float) {
         this.x = x;
         this.y = y;
     }
